+++ src/ByteDev.Configuration.Environment/EnvironmentVariableProvider.cs
@@ -224,6 +224,80 @@ namespace ByteDev.Configuration.Environment
         }
 
         /// <summary>
+        /// Retrieve an environment variable as a double. If it does not exist or it's value cannot be cast
+        /// then an exception will be thrown.
+        /// </summary>
+        /// <param name="name">Name of environment variable.</param>
+        /// <returns>Environment variable's value.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
+        /// <exception cref="T:ByteDev.Configuration.Environment.EnvironmentVariableNotExistException">Environment variable does not exist.</exception>
+        /// <exception cref="T:ByteDev.Configuration.Environment.UnexpectedEnvironmentVariableTypeException">Environment variable value is not a double.</exception>
+        public double GetDouble(string name)
+        {
+            var value = GetString(name);
+
+            if (double.TryParse(value, out var result))
+                return result;
+
+            throw new UnexpectedEnvironmentVariableTypeException(name, value, typeof(double));
+        }
+
+        /// <summary>
+        /// Retrieve an environment variable as a double. If it does not exist or it's value cannot be cast 
+        /// then the <paramref name="defaultValue" /> will be returned.
+        /// </summary>
+        /// <param name="name">Name of environment variable.</param>
+        /// <param name="defaultValue">Value to return if the environment variable does not exist.</param>
+        /// <returns>Environment variable's value.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
+        public double GetDoubleOrDefault(string name, double defaultValue = 0)
+        {
+            var value = GetStringOrDefault(name);
+
+            if (double.TryParse(value, out var result))
+                return result;
+
+            return defaultValue;
+        }
+
+        /// <summary>
+        /// Retrieve an environment variable as a decimal. If it does not exist or it's value cannot be cast
+        /// then an exception will be thrown.
+        /// </summary>
+        /// <param name="name">Name of environment variable.</param>
+        /// <returns>Environment variable's value.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
+        /// <exception cref="T:ByteDev.Configuration.Environment.EnvironmentVariableNotExistException">Environment variable does not exist.</exception>
+        /// <exception cref="T:ByteDev.Configuration.Environment.UnexpectedEnvironmentVariableTypeException">Environment variable value is not a decimal.</exception>
+        public decimal GetDecimal(string name)
+        {
+            var value = GetString(name);
+
+            if (decimal.TryParse(value, out var result))
+                return result;
+
+            throw new UnexpectedEnvironmentVariableTypeException(name, value, typeof(decimal));
+        }
+
+        /// <summary>
+        /// Retrieve an environment variable as a decimal. If it does not exist or it's value cannot be cast 
+        /// then the <paramref name="defaultValue" /> will be returned.
+        /// </summary>
+        /// <param name="name">Name of environment variable.</param>
+        /// <param name="defaultValue">Value to return if the environment variable does not exist.</param>
+        /// <returns>Environment variable's value.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
+        public decimal GetDecimalOrDefault(string name, decimal defaultValue = 0)
+        {
+            var value = GetStringOrDefault(name);
+
+            if (decimal.TryParse(value, out var result))
+                return result;
+
+            return defaultValue;
+        }
+
+        /// <summary>
         /// Retrieve an environment variable as a Uri. If it does not exist or it's value cannot be cast
         /// then an exception will be thrown.
         /// </summary>