+++ src/ByteDev.Configuration.Environment/IEnvironmentVariableProvider.cs
@@ -113,6 +113,27 @@ namespace ByteDev.Configuration.Environment
         long GetLongOrDefault(string name, long defaultValue = 0);
 
         /// <summary>
+        /// Retrieve an environment variable as a double. If it does not exist or it's value cannot be cast
+        /// then an exception will be thrown.
+        /// </summary>
+        /// <param name="name">Name of environment variable.</param>
+        /// <returns>Environment variable's value.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
+        /// <exception cref="T:ByteDev.Configuration.Environment.EnvironmentVariableNotExistException">Environment variable does not exist.</exception>
+        /// <exception cref="T:ByteDev.Configuration.Environment.UnexpectedEnvironmentVariableTypeException">Environment variable value is not a double.</exception>
+        double GetDouble(string name);
+
+        /// <summary>
+        /// Retrieve an environment variable as a double. If it does not exist or it's value cannot be cast 
+        /// then the <paramref name="defaultValue" /> will be returned.
+        /// </summary>
+        /// <param name="name">Name of environment variable.</param>
+        /// <param name="defaultValue">Value to return if the environment variable does not exist.</param>
+        /// <returns>Environment variable's value.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
+        double GetDoubleOrDefault(string name, double defaultValue = 0);
+
+        /// <summary>
         /// Retrieve an environment variable as a Uri. If it does not exist or it's value cannot be cast
         /// then an exception will be thrown.
         /// </summary>